@@ -36,47 +36,46 @@
     "test:cov": "jest --coverage"
   },
   "peerDependencies": {
-    "@nestjs/common": "7 - 9",
-    "@nestjs/core": "7 - 9"
+    "@nestjs/common": "7 - 10",
+    "@nestjs/core": "7 - 10"
   },
   "dependencies": {
-    "ssh2-sftp-client": "^8.1.0",
-    "ssh2-streams": "^0.4.10"
+    "ssh2-sftp-client": "^10.0.3"
   },
   "devDependencies": {
-    "@nestjs/cli": "^9.0.0",
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/config": "^2.2.0",
-    "@nestjs/core": "^9.0.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.0.0",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@types/jest": "28.1.8",
-    "@types/node": "^16.0.0",
-    "@types/ssh2-sftp-client": "^7.0.1",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "@nestjs/cli": "^10.0.0",
+    "@nestjs/common": "^10.0.0",
+    "@nestjs/config": "^3.1.1",
+    "@nestjs/core": "^10.0.0",
+    "@nestjs/schematics": "^10.0.0",
+    "@nestjs/testing": "^10.0.0",
+    "@semantic-release/changelog": "^6.0.3",
+    "@semantic-release/git": "^10.0.1",
+    "@types/jest": "^29.5.6",
+    "@types/node": "^20.8.9",
+    "@types/ssh2-sftp-client": "^7.1.4",
+    "@types/supertest": "^2.0.15",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
+    "eslint": "^8.52.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "husky": "^4.2.5",
     "import-sort-style-eslint": "^6.0.0",
-    "jest": "28.1.3",
+    "jest": "^29.7.0",
     "lint-staged": "^10.2.7",
-    "prettier": "^2.3.2",
+    "prettier": "^3.0.3",
     "prettier-plugin-import-sort": "0.0.4",
     "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
+    "rimraf": "^5.0.5",
     "rxjs": "^7.2.0",
-    "semantic-release": "^19.0.5",
-    "supertest": "^6.1.3",
-    "ts-jest": "28.0.8",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "4.1.0",
-    "typescript": "^4.7.4"
+    "semantic-release": "^22.0.5",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.5.0",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.2.2"
   },
   "release": {
     "branch": "master",
