@@ -4,7 +4,7 @@ import { DynamicModule, Global, Module } from '@nestjs/common';
 import { ConnectConfig } from 'ssh2';
 import { SftpClientService } from './sftp-client/sftp-client.service';
 
-import SftpClient = require('ssh2-sftp-client');
+import * as SftpClient from 'ssh2-sftp-client';
 
 @Global()
 @Module({
